This commit contains the proper feature calculation code including the management of the data folder in the feature directory
