--- .gitattributes
@@ -1,1 +1,0 @@
-*.csv filter=lfs diff=lfs merge=lfs -text
