--- data/test_data.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a3903d4ee05f140cbc567a8f42fc481b7a15777bcf1556a808cf54d190f83125
-size 9850304
